05_logging_monitoring/func05_integration/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
opencensus-extension-azure-functions
opencensus-ext-postgresql
opencensus-ext-sqlalchemy
redis
